pyproject.toml
[tool.poetry]
name = "zzsblog.github.io"
version = "0.1.0"
description = "zzsblog"
authors = ["JamzumSum <zzzzss990315@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
myst-parser = "^0.17.0"
pytz = "^2022.1"
pyyaml = "^6.0"
# extras: dev
pre-commit = {version = "^2.17.0", optional = true}
black = {version = "^22.1.0", optional = true}
isort = {version = "^5.10.1", optional = true}

[tool.poetry.dev-dependencies]
Sphinx = "^4.4.0"
sphinx-press-theme = "^0.8.0"

[tool.poetry.extras]
dev = [
    "pre-commit",
    "black",
    "isort"
]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 99

[tool.black]
line-length = 99
target-version = ['py37']
